Create validate sample sheet api command
